Upgrade matrix multiplication funciton
The used function now uses both OpenMP to parallelize the triple loop,
but also uses the transpose of B. This is done so that the lookup of
the values are cache-friendly, see
[this](http://stackoverflow.com/questions/22634121/openmp-c-matrix-multi
plication-run-slower-in-parallel).

The latter gives a small boost, it takes some time to transpose B, but
it’s still faster to to so.