Working copy of limited Cannon complete
This uses MPI's cartesian topology to shift things around.
Currently works only for num_procs = n^2 + 1, where n is the
dimension of the A and B matrices. Only square matrices, and
the sqrt(num_procs-1) % n == 0.

Next should be to try and extend this to where A is (m x l), B
is (l x n), where m, l, n does not need to be the same, and
may not be divisible by sqrt(num_procs-1).